Create text package on compose.components.text
