+++ WordPress/src/main/java/org/wordpress/android/ui/compose/components/text/Message.kt
@@ -1,17 +1,19 @@
-package org.wordpress.android.ui.compose.components
+package org.wordpress.android.ui.compose.components.text
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import org.wordpress.android.R.color
 import org.wordpress.android.R.dimen
 
 @Composable
-fun Subtitle(
+fun Message(
     text: String,
     modifier: Modifier = Modifier,
 ) {
@@ -19,8 +21,9 @@ fun Subtitle(
         text = text,
         fontSize = 17.sp,
         style = TextStyle(letterSpacing = (-0.01).sp),
+        color = colorResource(color.gray_50),
         modifier = modifier
-            .padding(horizontal = 30.dp)
-            .padding(top = 20.dp)
+            .padding(horizontal = dimensionResource(dimen.jp_migration_padding_horizontal))
+            .padding(top = 20.dp, bottom = 30.dp)
     )
 }
